+++ R/badge.R
@@ -16,12 +16,31 @@ use_cran_badge <- function(base_path = ".") {
   invisible(TRUE)
 }
 
+#' Use a README badge
+#'
+#' @param badge_name Badge name. Used in error message and alt text
+#' @param href,src Badge link and image src
+#' @inheritParams use_template
+#' @export
 use_badge <- function(badge_name, href, src, base_path = ".") {
+  if (has_badge(href, base_path)) {
+    return(FALSE)
+  }
+
   img <- paste0("![", badge_name, "](", src, ")")
   link <- paste0("[", img, "](", href, ")")
 
-  todo(
-    "Add a ", badge_name, " badge by adding the following line to your README:",
-    link
-  )
+  todo("Add a ", badge_name, " badge by adding the following line to your README:")
+  code_block(link)
+}
+
+has_badge <- function(href, base_path = ".") {
+  readme_path <- file.path(base_path, "README.md")
+  if (!file.exists(readme_path)) {
+    return(FALSE)
+  }
+
+  readme <- readLines(readme_path)
+  any(grepl(href, readme, fixed = TRUE))
+
 }