+++ R/ci.R
@@ -14,9 +14,7 @@ NULL
 #' @export
 #' @rdname ci
 use_travis <- function(browse = interactive(), base_path = ".") {
-  if (!uses_github(base_path)) {
-    stop("You must use_github() before using travis", call. = FALSE)
-  }
+  check_uses_github(base_path)
 
   use_template(
     "travis.yml",
@@ -25,18 +23,34 @@ use_travis <- function(browse = interactive(), base_path = ".") {
     base_path = base_path
   )
 
-  gh <- gh::gh_tree_remote(base_path)
-  travis_url <- file.path("https://travis-ci.org", gh$username, "/", gh$repo)
-  travis_img <- paste0(travis_url, ".svg?branch=master")
-
-  use_badge("Travis build status", travis_url, travis_img, base_path = base_path)
-
-  todo("Turn on travis for your repo at ", travis_url)
-  if (browse) {
-    utils::browseURL(travis_url)
-  }
+  travis_badge(base_path = base_path)
+  travis_activate(browser, base_path = base_path)
 
   invisible(TRUE)
+}
+
+
+travis_info <- function(base_path = ".") {
+  gh <- gh::gh_tree_remote(base_path)
+
+  url <- file.path("https://travis-ci.org", gh$username, gh$repo)
+  img <- paste0(url, ".svg?branch=master")
+
+  list(url = url, img = img)
+}
+
+travis_badge <- function(base_path) {
+  travis <- travis_info(base_path)
+  use_badge("Travis build status", travis$url, travis$img, base_path = base_path)
+}
+
+travis_activate <- function(browse = interactive(), base_path = ".") {
+  travis <- travis_info(base_path)
+
+  todo("Turn on travis for your repo at ", travis$url)
+  if (browse) {
+    utils::browseURL(travis$url)
+  }
 }
 
 
@@ -54,21 +68,16 @@ use_coverage <- function(type = c("codecov", "coveralls"), base_path = ".") {
   if (!uses_travis(base_path)) {
     stop("You must use_travis() first", call. = FALSE)
   }
+  type <- match.arg(type)
 
   use_dependency("covr", "Suggests", base_path = base_path)
 
-  gh <- gh::gh_tree_remote(base_path)
-  type <- match.arg(type)
-
   switch(type,
     codecov = {
-      use_template("codecov.yml", "codecov.yml", ignore = TRUE, base_path = base_path)
-      use_badge("Coverage status",
-        paste0("https://codecov.io/github/", gh$username, "/", gh$repo, "?branch=master"),
-        paste0("https://codecov.io/gh/", gh$username, "/", gh$repo, "/branch/master/graph/badge.svg")
-      )
-      todo("Add to `.travis.yml`:")
-      code(
+      use_template("codecov.yml", ignore = TRUE, base_path = base_path)
+      codecov_badge(base_path = base_path)
+      todo("Add to ", value(".travis.yml"), ":")
+      code_block(
         "after_success:",
         "  - Rscript -e 'covr::codecov()'"
       )
@@ -76,18 +85,32 @@ use_coverage <- function(type = c("codecov", "coveralls"), base_path = ".") {
 
     coveralls = {
       todo("Turn on coveralls for this repo at https://coveralls.io/repos/new")
-      use_badge("Coverage status",
-        paste0("https://coveralls.io/r/", gh$username, "/", gh$repo, "?branch=master"),
-        paste0("https://img.shields.io/coveralls/", gh$username, "/", gh$repo, ".svg")
-      )
-      todo("Add to `.travis.yml`")
-      code(
+      coveralls_badge(base_path = base_path)
+      todo("Add to ", value(".travis.yml"), ":")
+      code_block(
         "after_success:",
         "  - Rscript -e 'covr::coveralls()'"
       )
     })
 
   invisible(TRUE)
+}
+
+codecov_badge <- function(base_path = ".") {
+  gh <- gh::gh_tree_remote(base_path)
+
+  use_badge("Coverage status",
+    paste0("https://codecov.io/github/", gh$username, "/", gh$repo, "?branch=master"),
+    paste0("https://codecov.io/gh/", gh$username, "/", gh$repo, "/branch/master/graph/badge.svg")
+  )
+}
+
+coveralls_badge <- function(base_path = ".") {
+  gh <- gh::gh_tree_remote(base_path)
+  use_badge("Coverage status",
+    paste0("https://coveralls.io/r/", gh$username, "/", gh$repo, "?branch=master"),
+    paste0("https://img.shields.io/coveralls/", gh$username, "/", gh$repo, ".svg")
+  )
 }
 
 #' @rdname ci
@@ -101,7 +124,7 @@ use_appveyor <- function(base_path = ".") {
   gh <- gh::gh_tree_remote(base_path)
   todo("Turn on AppVeyor for this repo at https://ci.appveyor.com/projects\n")
   todo("Add an AppVeyor shield to your README.md:")
-  code(paste0(
+  code_block(paste0(
     "[![AppVeyor Build Status]",
     "(https://ci.appveyor.com/api/projects/status/github/", gh$username, "/", gh$repo, "?branch=master&svg=true)]",
     "(https://ci.appveyor.com/project/", gh$username, "/", gh$repo, ")"