+++ R/helpers.R
@@ -6,6 +6,7 @@
 #' @param ignore Should the newly created file be added to \code{.Rbuildignore?}
 #' @param open Should the new created file be opened in RStudio?
 #' @param base_path Path to package root.
+#' @return A logical vector indicating if file was modified.
 #' @keywords internal
 use_template <- function(template,
                          save_as = template,
@@ -16,18 +17,18 @@ use_template <- function(template,
                          ) {
 
   template_contents <- render_template(template, data)
-  write_over(base_path, save_as, template_contents)
+  new <- write_over(base_path, save_as, template_contents)
 
   if (ignore) {
     use_build_ignore(save_as, base_path = base_path)
   }
 
   if (open) {
-    todo(paste0("Modify ", value(save_as)))
+    todo("Modify ", value(save_as))
     open_in_rstudio(save_as, base_path = base_path)
   }
 
-  invisible(TRUE)
+  invisible(new)
 }
 
 render_template <- function(template, data = list()) {
@@ -69,16 +70,16 @@ use_description_field <- function(name, value, base_path = ".", overwrite = FALS
     return()
 
   if (is.na(curr) || overwrite) {
-    done(paste0("Setting DESCRIPTION ", field(name), " to ", value(value)))
+    done("Setting ", field(name), " field in DESCRIPTION to ", value(value))
     desc::desc_set(name, value, file = base_path)
   }
 }
 
-use_dependency <- function(package, type, base_path = ".") {
+use_dependency <- function(package, type, version = "*", base_path = ".") {
   stopifnot(is.character(package), length(package) == 1)
   stopifnot(is.character(type), length(type) == 1)
 
-  if (!requireNamespace(package, quietly = TRUE)) {
+  if (package != "R" && !requireNamespace(package, quietly = TRUE)) {
     stop(package, " must be installed before you can take a dependency on it",
       call. = FALSE)
   }
@@ -88,10 +89,17 @@ use_dependency <- function(package, type, base_path = ".") {
   type <- types[[match.arg(tolower(type), names(types))]]
 
   deps <- desc::desc_get_deps(base_path)
-  has_dep <- any(deps$package == package & deps$type == type)
-  if (!has_dep) {
-    done(paste0("Adding ", value(package), " to DESCRIPTION ", field(type)))
-    desc::desc_set_dep(package, type, file = file.path(base_path, "DESCRIPTION"))
+
+  matching_dep <- deps$package == package & deps$type == type
+  to_add <- !any(matching_dep)
+  to_set <- any(matching_dep & deps$version != version)
+
+  if (to_add) {
+    done("Adding ", value(package), " to ", field(type), " field in DESCRIPTION")
+    desc::desc_set_dep(package, type, version = version, file = base_path)
+  } else if (to_set) {
+    done("Setting ", value(package), " version to ", field(version), " field in DESCRIPTION")
+    desc::desc_set_dep(package, type, version = version, file = base_path)
   }
 
   invisible()
@@ -111,7 +119,7 @@ use_directory <- function(path,
       stop(value(path), " exists but is not a directory.", call. = FALSE)
     }
   } else {
-    done(paste0("Creating ", value(path, "/")))
+    done("Creating ", value(path, "/"))
     ok <- dir.create(pkg_path, showWarnings = FALSE, recursive = TRUE)
 
     if (!ok) {