+++ R/style.R
@@ -1,21 +1,20 @@
 bullet <- function(lines, bullet) {
   lines <- paste0(bullet, " ", crayon::black(lines))
-  cat_line(paste0(lines, "\n"))
+  message(lines)
 }
 
 todo_bullet <- function() crayon::red(clisymbols::symbol$bullet)
 
 todo <- function(...) {
-  bullet(c(...), bullet = todo_bullet())
+  bullet(paste0(...), bullet = todo_bullet())
 }
 done <- function(...) {
-  bullet(c(...), bullet = crayon::green(clisymbols::symbol$tick))
+  bullet(paste0(...), bullet = crayon::green(clisymbols::symbol$tick))
 }
 
-code <- function(...) {
-  lines <- c(...)
-  block <- paste0("  ", lines, "\n", collapse = "")
-  cat_line(crayon::make_style("darkgrey")(block))
+code_block <- function(...) {
+  block <- paste0("  ", c(...), collapse = "\n")
+  message(crayon::make_style("darkgrey")(block))
 }
 
 cat_line <- function(...) {
@@ -29,4 +28,13 @@ field <- function(...) {
 value <- function(...) {
   x <- paste0(...)
   crayon::blue(encodeString(x, quote = "'"))
+}
+
+code <- function(...) {
+  x <- paste0(...)
+  crayon::make_style("darkgrey")(encodeString(x, quote = "`"))
+}
+
+collapse <- function(x, sep = ", ") {
+  paste0(x, collapse = sep)
 }