+++ NEWS.md
@@ -1,5 +1,20 @@
 # usethis 0.0.0.9000
 
+* New `use_tidy_description()` puts description fields in a standard order
+  and alphabetises dependencies.
+
+* New `use_pipe()` creates a template to use magrittr's `%>%` in your package
+  (#15).
+
+* `create_package()` no longer creates a dependency on the current version of 
+  R.
+
+* New `use_tidy_ci()` which sets up travis and codecov using the tidyverse
+  conventions (#14)
+
+* New `use_badge()` for adding any badge to a README. Now only prints a 
+  todo message if the badge does not already exist.
+
 * `use_revdep()` creates structure for use with revdepcheck package, the
   preferred way to run revdepchecks. (#33)